+++ assets/src/components/cd/pipelines/PipelineDetails.tsx
