+++ assets/src/components/cd/pipelines/job/PipelineJob.tsx
@@ -41,7 +41,7 @@ import { Body2P, Subtitle2H1 } from 'components/utils/typography/Text'
 
 import { POLL_INTERVAL } from 'components/cd/ContinuousDeployment'
 
-import { PIPELINES_CRUMBS } from '../Pipelines'
+import { PIPELINES_CRUMBS } from '../PipelineDetails'
 
 const getDirectory = () => [
   { path: 'logs', label: 'Logs', enabled: true },
